pci: extend MSI interrupt vector number to 32 bits
Also increase MSI message data size to 32 bits according to PCIe spec.

Remove 0xff check for MSI interrupt because it is potentially valid interrupt vector number. Reject 0xff only for legacy pin interrupts.

- MSI-X supports up to 2048 interrupts that do not fit to `uint8`.

- Non-x86 systems may use separate interrupt vector ranges for
hard-wired interrupts and MSI interrupts so `uint8` is not enough to
represent all of them.

+++ src/system/kernel/arch/x86/msi.cpp
@@ -14,7 +14,7 @@
 
 struct MSIConfiguration {
 	uint64*	fAddress;
-	uint16* fData;
+	uint32* fData;
 };
 
 static MSIConfiguration sMSIConfigurations[NUM_IO_VECTORS];
@@ -45,8 +45,8 @@ msi_supported()
 
 
 status_t
-msi_allocate_vectors(uint8 count, uint8 *startVector, uint64 *address,
-	uint16 *data)
+msi_allocate_vectors(uint32 count, uint32 *startVector, uint64 *address,
+	uint32 *data)
 {
 	if (!sMSISupported)
 		return B_UNSUPPORTED;
@@ -66,7 +66,7 @@ msi_allocate_vectors(uint8 count, uint8 *startVector, uint64 *address,
 	sMSIConfigurations[vector].fData = data;
 	x86_set_irq_source(vector, IRQ_SOURCE_MSI);
 
-	*startVector = (uint8)vector;
+	*startVector = (uint32)vector;
 	*address = MSI_ADDRESS_BASE | (sBootCPUAPICId << MSI_DESTINATION_ID_SHIFT)
 		| MSI_NO_REDIRECTION | MSI_DESTINATION_MODE_PHYSICAL;
 	*data = MSI_TRIGGER_MODE_EDGE | MSI_DELIVERY_MODE_FIXED
@@ -79,7 +79,7 @@ msi_allocate_vectors(uint8 count, uint8 *startVector, uint64 *address,
 
 
 void
-msi_free_vectors(uint8 count, uint8 startVector)
+msi_free_vectors(uint32 count, uint32 startVector)
 {
 	if (!sMSISupported) {
 		panic("trying to free msi vectors but msi not supported\n");
@@ -94,7 +94,7 @@ msi_free_vectors(uint8 count, uint8 startVector)
 
 
 void
-msi_assign_interrupt_to_cpu(uint8 irq, int32 cpu)
+msi_assign_interrupt_to_cpu(uint32 irq, int32 cpu)
 {
 	uint32 apic_id = x86_get_cpu_apic_id(cpu);
